Implement a simple Trie
Implement a simple prefix tree class, or Trie. Support inserting and removing
words, and iterating over all complete words with a given prefix. Nodes of the
Trie are themselves Trie objects. Support arbitrary nodes, not just characters,
but provide a subclass, StringTrie, that behaves as expected for the usual case
of storing sequences of characters