Move parts of Trie into inner classes
Define separate Node and View classes within Trie, and offload some
functionality and data storage to these. Previously, nodes of the trie were
themselves Trie objects. This meant that calling 'remove' on a Trie
corresponding to a descendent node could leave an unpruned branch of the
original Trie. By separating the node class, this can no longer happen; also,
it will be possible to safely cache len(Trie) in a further update. The main
instance where a sub-Trie was accessed was via the Trie.__getitem__
method. This method now instead returns a View object, which allows read-only
access to the sub-Trie for ths same reasons as above.